Update README.md
12th Arts, Commerce, Science Exam in 2024

The board exam authority is diligently working to upload the date sheet for the 12th class final exams as soon as possible. Students can access the date sheet on the official website of their respective state boards once it is uploaded. Additionally, we will also update the 12th Time Table 2024 on our website, so it is advisable to stay connected with us to receive the latest information regarding the date sheet.

For students who are planning to appear in the Intermediate Board Exams, it is crucial to regularly check the official website of their respective boards for the 12th Class Exam Date Sheet. In addition, we have provided helpful links on our website to provide information about HSC and Intermediate Time Table 2024. These links can serve as valuable resources for students seeking information and updates related to the exam schedule.